+++ test/Eawv.Service.Integration.Tests/ListTests/CreateListTest.cs
@@ -17,7 +17,11 @@ namespace Eawv.Service.Integration.Tests.ListTests;
 
 public class CreateListTest : BaseRestTest
 {
-    private static readonly string Url = $"api/elections/{ElectionMockData.ProporzElection.Id}/lists";
+    private static readonly string UrlProporzElection = $"api/elections/{ElectionMockData.ProporzElection.Id}/lists";
+    private static readonly string UrlFutureElection = $"api/elections/{ElectionMockData.FutureUnavailableElection.Id}/lists";
+    private static readonly string UrlFutureAvailableElection = $"api/elections/{ElectionMockData.FutureAvailableElection.Id}/lists";
+    private static readonly string UrlPastElection = $"api/elections/{ElectionMockData.PastElection.Id}/lists";
+    private static readonly string UrlArchivedElection = $"api/elections/{ElectionMockData.ArchivedElection.Id}/lists";
 
     public CreateListTest(TestApplicationFactory factory)
         : base(factory)
@@ -33,14 +37,14 @@ public class CreateListTest : BaseRestTest
     [Fact]
     public async Task TestAsElectionAdmin()
     {
-        var list = await GetSuccessfulResponse<ListModel>(() => ElectionAdminClient.PostAsJsonAsync(Url, NewValidRequest()));
+        var list = await GetSuccessfulResponse<ListModel>(() => ElectionAdminClient.PostAsJsonAsync(UrlProporzElection, NewValidRequest()));
         list.MatchSnapshot(x => x.Id);
     }
 
     [Fact]
     public async Task TestAsUser()
     {
-        var list = await GetSuccessfulResponse<ListModel>(() => UserClient.PostAsJsonAsync(Url, NewValidRequest()));
+        var list = await GetSuccessfulResponse<ListModel>(() => UserClient.PostAsJsonAsync(UrlProporzElection, NewValidRequest()));
         list.MatchSnapshot(x => x.Id);
     }
 
@@ -48,7 +52,7 @@ public class CreateListTest : BaseRestTest
     public async Task TestWithRepresentativeOfDifferentTenantShouldThrow()
     {
         await AssertStatus(
-            () => UserClient.PostAsJsonAsync(Url, NewValidRequest(x => x.Representative = UserMockData.SpUser.Id)),
+            () => UserClient.PostAsJsonAsync(UrlProporzElection, NewValidRequest(x => x.Representative = UserMockData.SpUser.Id)),
             HttpStatusCode.Forbidden);
     }
 
@@ -56,7 +60,7 @@ public class CreateListTest : BaseRestTest
     public async Task TestWithDeputiesOfDifferentTenantShouldThrow()
     {
         await AssertStatus(
-            () => UserClient.PostAsJsonAsync(Url, NewValidRequest(x => x.DeputyUsers = [UserMockData.SpUser.Id])),
+            () => UserClient.PostAsJsonAsync(UrlProporzElection, NewValidRequest(x => x.DeputyUsers = [UserMockData.SpUser.Id])),
             HttpStatusCode.Forbidden);
     }
 
@@ -64,8 +68,59 @@ public class CreateListTest : BaseRestTest
     public async Task TestWithMembersOfDifferentTenantShouldThrow()
     {
         await AssertStatus(
-            () => UserClient.PostAsJsonAsync(Url, NewValidRequest(x => x.MemberUsers = [UserMockData.SpUser.Id])),
+            () => UserClient.PostAsJsonAsync(UrlProporzElection, NewValidRequest(x => x.MemberUsers = [UserMockData.SpUser.Id])),
             HttpStatusCode.Forbidden);
+    }
+
+    /// <summary>
+    /// Ensures that creating a list for a future election that is already visible for parties returns Forbidden.
+    /// Lists cannot be created outside the submission time period.
+    /// </summary>
+    /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
+    [Fact]
+    public async Task TestWithFutureAvailableElectionShouldThrow()
+    {
+        await AssertStatus(
+            () => UserClient.PostAsJsonAsync(UrlFutureAvailableElection, NewValidRequest()),
+            HttpStatusCode.Forbidden);
+    }
+
+    /// <summary>
+    /// Ensures that creating a list for a future election that is not yet visible for parties returns NotFound.
+    /// Lists cannot be seen or created before the election becomes available.
+    /// </summary>
+    /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
+    [Fact]
+    public async Task TestWithFutureElectionShouldThrow()
+    {
+        await AssertStatus(
+            () => UserClient.PostAsJsonAsync(UrlFutureElection, NewValidRequest()),
+            HttpStatusCode.NotFound);
+    }
+
+    /// <summary>
+    /// Ensures that creating a list for a past election that is not yet archieved returns Forbidden.
+    /// Lists cannot be created outside the submission time period.
+    /// </summary>
+    /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
+    [Fact]
+    public async Task TestWithPastElectionShouldThrow()
+    {
+        await AssertStatus(
+            () => UserClient.PostAsJsonAsync(UrlPastElection, NewValidRequest()),
+            HttpStatusCode.Forbidden);
+    }
+
+    /// <summary>
+    /// Ensures that creating a list for an archived election returns BadRequest.
+    /// </summary>
+    /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
+    [Fact]
+    public async Task TestWithArchivedElectionShouldThrow()
+    {
+        await AssertStatus(
+            () => UserClient.PostAsJsonAsync(UrlArchivedElection, NewValidRequest()),
+            HttpStatusCode.BadRequest);
     }
 
     protected override IEnumerable<string> AuthorizedRoles()
@@ -76,7 +131,7 @@ public class CreateListTest : BaseRestTest
 
     protected override Task<HttpResponseMessage> AuthorizationTestCall(HttpClient httpClient)
     {
-        return httpClient.PostAsJsonAsync(Url, NewValidRequest());
+        return httpClient.PostAsJsonAsync(UrlProporzElection, NewValidRequest());
     }
 
     private ModifyListModel NewValidRequest(Action<ModifyListModel> customizer = null)